Move AuthToken bean into api.common bundle
